% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rjsize_i.R
\name{crjsize_i}
\alias{crjsize_i}
\alias{rjsize_i}
\title{Update \eqn{J_i}}
\usage{
crjsize_i(jtime_i, vim1, vi, yi, mu, k, theta, sigma_v, rho, lmbd, sigma_j, h)

rjsize_i(jtime_i, vim1, vi, yi, mu, k, theta, sigma_v, rho, lmbd, sigma_j, h)
}
\arguments{
\item{jtime_i}{jump indicator at epoch i.}

\item{vim1}{variance \eqn{v_{i-1}}.}

\item{vi}{variance \eqn{v_i}.}

\item{yi}{return \eqn{y_i}.}

\item{mu}{parameter \eqn{\mu}.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma_v}{parameter \eqn{\sigma_v}.}

\item{rho}{parameter \eqn{\rho}.}

\item{lmbd}{parameter \eqn{\lambda}.}

\item{sigma_j}{paramter \eqn{\sigma_j}.}

\item{h}{time unit.}
}
\value{
new value of \eqn{J_i}, a scale value.
}
\description{
Generate new jump size.
}
