% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rjtime_i.R
\name{crjtime_i}
\alias{crjtime_i}
\alias{rjtime_i}
\title{Update \eqn{I_i}}
\usage{
crjtime_i(jsize_i, vim1, vi, yi, mu, k, theta, sigma_v, rho, lmbd, sigma_j, h)

rjtime_i(jsize_i, vim1, vi, yi, mu, k, theta, sigma_v, rho, lmbd, sigma_j, h)
}
\arguments{
\item{jsize_i}{jump size at epoch i.}

\item{vim1}{variance \eqn{v_{i-1}}.}

\item{vi}{variance \eqn{v_i}.}

\item{yi}{return \eqn{y_i}.}

\item{mu}{parameter \eqn{\mu}.}

\item{k}{parameter \eqn{k}.}

\item{theta}{parameter \eqn{\theta}.}

\item{sigma_v}{parameter \eqn{\sigma_v}.}

\item{rho}{parameter \eqn{\rho}.}

\item{lmbd}{parameter \eqn{\lambda}.}

\item{sigma_j}{paramter \eqn{\sigma_j}.}

\item{h}{time unit.}
}
\value{
new value of \eqn{J_i}, a scale value.
}
\description{
Generate new jump indicator, whether there is a jump or not.
}
