% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rlambda.R
\name{crlambda}
\alias{crlambda}
\alias{rlambda}
\title{Update \eqn{\lambda}}
\usage{
crlambda(prior_a, prior_b, jtime)

rlambda(prior_a, prior_b, jtime)
}
\arguments{
\item{prior_a}{the first conjugate prior shape}

\item{prior_b}{the second conjugate prior shape}

\item{jtime}{vector of jump indicators \[\eqn{I_1, ..., I_N}\], I_n = 1 or 0.}
}
\value{
new value for \eqn{\lambda}, a scale value.
}
\description{
Parameter \eqn{\lambda} has a Beta posterior if its prior is also a Beta.
}
\examples{
prior_a = 2
prior_b = 40
jtime = rbinom(4, size=1, prob=0.01)
rlambda(prior_a, prior_b, jtime)
}
