% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/rsigma_j.R
\name{crsigma_j}
\alias{crsigma_j}
\alias{rsigma_j}
\title{Update \eqn{\sigma_y^2}}
\usage{
crsigma_j(prior_shape, prior_scale, jsize)

rsigma_j(prior_shape, prior_scale, jsize)
}
\arguments{
\item{prior_shape}{conjugate prior shape parameter}

\item{prior_scale}{conjugate prior scale parameter}

\item{jsize}{vector of jump sizes \[\eqn{J_1, ..., J_N}\]}
}
\value{
new value for \eqn{\sigma_y}, a scale value.
}
\description{
Parameter \eqn{\sigma_y^2} has an Inverse Gamma posterior if its prior is
also an Inverse Gamma.
}
\examples{
prior_shape = 5.0
prior_scale = 20
jsize = rnorm(4, mean=0, sd=0.5)
rsigma_j(prior_shape, prior_scale, jsize)
}
